refactor(ui): split Canvas annotation helpers
